Move hashring to lib/datastructures, default to XXH3
Creates a new lib/datastructures module so we can pull in
github.com/zeebo/xxh3 without polluting the stdlib-only lib/std.
The new module is the home for hashring (today) and future
hash-aware data structures (planned).

XXH3 replaces FNV-1a as the default hash. FNV-1a's weak avalanche
caused virtual-node clustering on near-sequential inputs (e.g.
consecutive IPs): with 100 members and 1024 IPs the busiest member
owned ~33 keys (mean 10.24) and 2 members owned zero. XXH3 reaches
the Poisson sampling floor of uniform distribution. WithHash still
lets callers swap in any deterministic byte-slice hasher.
