+++ lib/datastructures/go.mod
@@ -1,0 +1,10 @@
+module github.com/projectcalico/calico/lib/datastructures
+
+go 1.26.3
+
+require github.com/zeebo/xxh3 v1.1.0
+
+require (
+	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
+	golang.org/x/sys v0.30.0 // indirect
+)
